Fix unsatable mv sql tests
