+++ fe/fe-core/src/test/java/com/starrocks/catalog/MaterializedViewTest.java
@@ -18,6 +18,7 @@ package com.starrocks.catalog;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.starrocks.alter.AlterJobV2;
 import com.starrocks.analysis.Expr;
 import com.starrocks.analysis.FunctionCallExpr;
 import com.starrocks.analysis.SlotRef;
@@ -39,6 +40,7 @@ import com.starrocks.sql.analyzer.SemanticException;
 import com.starrocks.sql.ast.AlterTableStmt;
 import com.starrocks.sql.ast.ShowCreateTableStmt;
 import com.starrocks.sql.ast.StatementBase;
+import com.starrocks.sql.optimizer.MVTestUtils;
 import com.starrocks.sql.parser.SqlParser;
 import com.starrocks.thrift.TStorageMedium;
 import com.starrocks.thrift.TStorageType;
@@ -69,13 +71,13 @@ public class MaterializedViewTest {
 
     @Before
     public void setUp() {
-        UtFrameUtils.createMinStarRocksCluster();
-
         connectContext = UtFrameUtils.createDefaultCtx();
         starRocksAssert = new StarRocksAssert(connectContext);
 
         // set default config for async mvs
         UtFrameUtils.setDefaultConfigForAsyncMVTest(connectContext);
+
+        UtFrameUtils.createMinStarRocksCluster();
 
         columns.add(new Column("k1", ScalarType.createType(PrimitiveType.TINYINT), true, null, "", ""));
         columns.add(new Column("k2", ScalarType.createType(PrimitiveType.SMALLINT), true, null, "", ""));
@@ -612,16 +614,24 @@ public class MaterializedViewTest {
                 .withMaterializedView("create materialized view index_mv_to_check\n" +
                         "distributed by hash(k2) buckets 3\n" +
                         "as select k2, sum(v1) as total from table1 group by k2;");
-        String bitmapSql = "create index index1 ON test.index_mv_to_check (k2) USING BITMAP COMMENT 'balabala'";
-        String bloomfilterSql = "alter table test.index_mv_to_check set (\"bloom_filter_columns\"=\"k2\")";
 
+        Database db = connectContext.getGlobalStateMgr().getDb("test");
+        Assert.assertNotNull(db);
+        Table table = db.getTable("index_mv_to_check");
+        Assert.assertNotNull(table);
+
+        String bitmapSql = "create index index1 ON test.index_mv_to_check (k2) USING BITMAP COMMENT 'balabala'";
         AlterTableStmt alterMVStmt = (AlterTableStmt) UtFrameUtils.parseStmtWithNewParser(bitmapSql, connectContext);
         DDLStmtExecutor.execute(alterMVStmt, connectContext);
-        waitForSchemaChangeAlterJobFinish();
+        Optional<AlterJobV2> job = MVTestUtils.findAlterJobV2(db.getId(), table.getId());
+        Assert.assertTrue("Alter job should be present", job.isPresent());
+        waitForSchemaChangeAlterJobFinish(job.get());
 
+        String bloomfilterSql = "alter table test.index_mv_to_check set (\"bloom_filter_columns\"=\"k2\")";
         alterMVStmt = (AlterTableStmt) UtFrameUtils.parseStmtWithNewParser(bloomfilterSql, connectContext);
         DDLStmtExecutor.execute(alterMVStmt, connectContext);
-        waitForSchemaChangeAlterJobFinish();
+        job = MVTestUtils.findAlterJobV2(db.getId(), table.getId());
+        waitForSchemaChangeAlterJobFinish(job.get());
 
         Assert.assertEquals(QueryState.MysqlStateType.OK, connectContext.getState().getStateType());
     }
