+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewritePreprocessorTest.java
@@ -345,8 +345,11 @@ public class MvRewritePreprocessorTest extends MvRewriteTestBase {
                 Assert.assertTrue(containsMV(relatedMVs, "mv_1", "mv_2", "mv_3", "mv_4"));
 
                 Set<MaterializedView> validMVs = preprocessor.chooseBestRelatedMVs(queryTables, relatedMVs, logicalTree);
+                // disable plan cache will make the test more stable
+                connectContext.getSessionVariable().setEnableMaterializedViewPlanCache(false);
                 Assert.assertEquals(2, validMVs.size());
                 Assert.assertTrue(containsMV(validMVs, "mv_1", "mv_3"));
+                connectContext.getSessionVariable().setEnableMaterializedViewPlanCache(true);
 
                 // if mv_3 is in the plan cache
                 MaterializedView mv3 = getMv("test", "mv_3");
@@ -394,6 +397,7 @@ public class MvRewritePreprocessorTest extends MvRewriteTestBase {
 
         int oldVal = connectContext.getSessionVariable().getCboMaterializedViewRewriteRelatedMVsLimit();
         connectContext.getSessionVariable().setCboMaterializedViewRewriteRelatedMVsLimit(2);
+
         starRocksAssert.withMaterializedViews(mvs, (obj) -> {
 
             long currentTime = System.currentTimeMillis();
@@ -417,6 +421,8 @@ public class MvRewritePreprocessorTest extends MvRewriteTestBase {
                 Assert.assertTrue(containsMV(relatedMVs, "mv_1", "mv_2", "mv_3", "mv_4"));
 
                 Set<MaterializedView> validMVs = preprocessor.chooseBestRelatedMVs(queryTables, relatedMVs, logicalTree);
+                // disable plan cache will make the test more stable
+                connectContext.getSessionVariable().setEnableMaterializedViewPlanCache(false);
                 Assert.assertEquals(2, validMVs.size());
                 Assert.assertTrue(containsMV(validMVs, "mv_1", "mv_3"));
 
@@ -424,6 +430,7 @@ public class MvRewritePreprocessorTest extends MvRewriteTestBase {
                 validMVs = preprocessor.chooseBestRelatedMVs(queryTables, relatedMVs, logicalTree);
                 Assert.assertEquals(1, validMVs.size());
                 Assert.assertTrue(containsMV(validMVs, "mv_3"));
+                connectContext.getSessionVariable().setEnableMaterializedViewPlanCache(true);
 
                 // if mv_3 is in the plan cache
                 connectContext.getSessionVariable().setCboMaterializedViewRewriteRelatedMVsLimit(2);
