--- test/sql/test_random_distribution/R/test_random_distribution_tablet_prune
@@ -1,32 +1,0 @@
--- name: test_random_distribution_tablet_prune @slow
-CREATE TABLE t1 (
-  num int,
-  dt date
-)
-DISTRIBUTED BY RANDOM 
-PROPERTIES(
-	'bucket_size' = '100'
-);
--- result:
--- !result
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
--- result:
--- !result
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
--- result:
--- !result
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
--- result:
--- !result
-function: wait_table_rowcount_not_empty("t1", 30)
--- result:
-None
--- !result
-select count(*) from t1 limit 1;
--- result:
-300
--- !result
-select * from t1 limit 1;
--- result:
-1	2020-06-18
--- !result
