--- test/sql/test_random_distribution/T/test_random_distribution_tablet_prune
@@ -1,19 +1,0 @@
--- name: test_random_distribution_tablet_prune @slow
-CREATE TABLE t1 (
-  num int,
-  dt date
-)
-DISTRIBUTED BY RANDOM 
-PROPERTIES(
-	'bucket_size' = '100'
-);
-
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
-insert into t1 select generate_series, "2020-06-18" from table(generate_series(1, 100));
-
--- wait until row counts are updated
-function: wait_table_rowcount_not_empty("t1", 30)
-
-select count(*) from t1 limit 1;
-select * from t1 limit 1;
